requirements.txt
cellpose==4.0.9
fastapi==0.135.1
gradio
onnxruntime==1.24.4
matplotlib==3.10.8
numpy==2.4.3
opencv-python-headless==4.11.0.86
Pillow==11.0.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydicom==3.0.1
python-multipart==0.0.22
reportlab==4.4.10
scikit-learn==1.8.0
scipy==1.17.1
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
timm==1.0.25
torch==2.10.0
torchvision==0.25.0
torchxrayvision==1.4.0
ultralytics==8.4.21
uvicorn==0.41.0